+++ pkg/workloads/docker/manager.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package containerd
+package docker
 
 import (
 	"github.com/cilium/cilium/api/v1/models"
